check for kubectl path before symlinking

+++ 01a-install_first_rke2_server.sh
@@ -206,8 +206,11 @@ copy_kubeconfig_file_and_kubectl() {
       cp /etc/rancher/${K8S_DISTRO}/${K8S_DISTRO}.yaml ~/.kube/config
       echo
 
-      echo "COMMAND: ln -s /var/lib/rancher/${K8S_DISTRO}/bin/kubectl /usr/local/bin/"
-      ln -s /var/lib/rancher/${K8S_DISTRO}/bin/kubectl /usr/local/bin/
+      if [ -e "/var/lib/rancher/${K8S_DISTRO}/bin/kubectl" ]
+      then
+        echo "COMMAND: ln -s /var/lib/rancher/${K8S_DISTRO}/bin/kubectl /usr/local/bin/"
+        ln -s /var/lib/rancher/${K8S_DISTRO}/bin/kubectl /usr/local/bin/
+      fi
       echo
 
       echo "COMMAND: kubectl completion bash | sudo tee /etc/bash_completion.d/kubectl > /dev/null"
