+++ 02c-symlink_kubectl.sh
@@ -19,10 +19,13 @@ then
   rm /usr/local/bin/kubectl
 fi
 
-echo
-echo "COMMAND: ln -s /var/lib/rancher/${K8S_DISTRO}/bin/kubectl /usr/local/bin/"
-ln -s /var/lib/rancher/${K8S_DISTRO}/bin/kubectl /usr/local/bin/
-echo
+if [ -e "/var/lib/rancher/${K8S_DISTRO}/bin/kubectl" ]
+then
+  echo
+  echo "COMMAND: ln -s /var/lib/rancher/${K8S_DISTRO}/bin/kubectl /usr/local/bin/"
+  ln -s /var/lib/rancher/${K8S_DISTRO}/bin/kubectl /usr/local/bin/
+  echo
+fi
 
 echo "COMMAND: kubectl completion bash | sudo tee /etc/bash_completion.d/kubectl > /dev/null"
 kubectl completion bash | sudo tee /etc/bash_completion.d/kubectl > /dev/null
